Refactored FromRecord a bit
